Extract Sidebar::Field and rename it to SidebarField
